Rename base/ to common/
