+++ src/common/data/EscapeSequences.ts
