+++ src/core/input/Keyboard.ts
@@ -4,9 +4,9 @@
  * @license MIT
  */
 
-import { IKeyboardEvent } from '../../base/Types';
+import { IKeyboardEvent } from '../../common/Types';
 import { IKeyboardResult, KeyboardResultType } from '../Types';
-import { C0 } from '../../base/data/EscapeSequences';
+import { C0 } from '../../common/data/EscapeSequences';
 
 // reg + shift key mappings for digits and special chars
 const KEYCODE_KEY_MAPPINGS: { [key: number]: [string, string]} = {
